+++ src/plugins/syslog/syslog_plugin.cpp
@@ -88,7 +88,7 @@ void SyslogPlugin::_begin()
         _port = cfg.getPort();
 
         if (config.hasZeroConf(_hostname)) {
-            config.resolveZeroConf(getFriendlyName(), _hostname, _port, [](const String &hostname, const IPAddress &address, uint16_t port, MDNSResolver::ResponseType type) {
+            config.resolveZeroConf(getFriendlyName(), _hostname, _port, [](const String &hostname, const IPAddress &address, uint16_t port, const String &resolved, MDNSResolver::ResponseType type) {
                 plugin._zeroConfCallback(hostname, address, port, type);
             });
         }